+++ src/main.rs
@@ -1,4 +1,5 @@
 use std::{error::Error, fs};
+use crate::app::App;
 use crate::model::model::{Config, StatefulList, StoredConnection};
 use crate::terminal::{init_error_hooks, init_terminal, restore_terminal};
 
@@ -6,12 +7,9 @@ mod utils;
 mod model;
 mod terminal;
 mod ui;
+mod app;
 
-struct App {
-    items: StatefulList,
-    new_item_popup: bool,
-    new_connection: StoredConnection,
-}
+
 
 fn main() -> Result<(), Box<dyn Error>> {
     let cfg: Config = confy::load_path("/home/yenqw/Code/ssh-manager-rust/config.toml")?;
